+++ src/main/java/land/leets/global/cron/SendMailCron.java
@@ -1,15 +1,11 @@
 package land.leets.global.cron;
 
-import java.util.List;
-
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import land.leets.domain.application.domain.Application;
-import land.leets.domain.application.domain.repository.ApplicationRepository;
 import land.leets.domain.application.type.ApplicationStatus;
 import land.leets.domain.mail.usecase.SendFinalMailImpl;
 import land.leets.domain.mail.usecase.SendPaperMailImpl;
-import org.springframework.scheduling.annotation.Scheduled;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -18,33 +14,30 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class SendMailCron {
 
-    private final ApplicationRepository applicationRepository;
-    private final SendFinalMailImpl sendFinalMailImpl;
-    private final SendPaperMailImpl sendPaperMailImpl;
+	private final SendFinalMailImpl sendFinalMailImpl;
+	private final SendPaperMailImpl sendPaperMailImpl;
 
-    @Scheduled(cron = "0 5 20 7 9 ?")
-    public void sendPaperMail() {
-        for (ApplicationStatus status : ApplicationStatus.papers()) {
-            List<Application> applications = applicationRepository.findAllByApplicationStatus(status);
-            sendPaperMailImpl.execute(status, applications);
-        }
+	@Scheduled(cron = "0 0 23 8 9 ?")
+	public void sendPaperMail() {
+		for (ApplicationStatus status : ApplicationStatus.papers()) {
+			sendPaperMailImpl.execute(status);
+		}
 
-        log.info("Send paper result mail successfully.");
-    }
+		log.info("Send paper result mail successfully.");
+	}
 
-    @Scheduled(cron = "0 0 10 12 9 ?")
-    public void sendFinalMail() {
-        for (ApplicationStatus status : ApplicationStatus.finals()) {
-            List<Application> applications = applicationRepository.findAllByApplicationStatus(status);
-            sendFinalMailImpl.execute(status, applications);
-        }
+	@Scheduled(cron = "0 0 10 12 9 ?")
+	public void sendFinalMail() {
+		for (ApplicationStatus status : ApplicationStatus.finals()) {
+			sendFinalMailImpl.execute(status);
+		}
 
-        log.info("Send final result mail successfully.");
-    }
+		log.info("Send final result mail successfully.");
+	}
 
-    //     @Scheduled(cron = "0 28 23 1 9 ?")
-    public void sendPlusMail() {
-        log.info("Send final result mail successfully.");
-    }
+	//     @Scheduled(cron = "0 28 23 1 9 ?")
+	public void sendPlusMail() {
+		log.info("Send final result mail successfully.");
+	}
 }
