+++ src/test/java/land/leets/TestFixture.java
@@ -8,6 +8,7 @@ import land.leets.domain.application.type.ApplicationStatus;
 import land.leets.domain.application.type.Position;
 import land.leets.domain.application.type.SubmitStatus;
 import land.leets.domain.user.domain.User;
+import land.leets.global.mail.dto.MailDto;
 
 public class TestFixture {
 
@@ -17,4 +18,6 @@ public class TestFixture {
 		"portfolio", Position.DEV, "c", "2024.06.02", "3시", "motive",
 		"schedule", "capability", "conflict", "passion", LocalDateTime.now(),
 		ApplicationStatus.PASS, SubmitStatus.SUBMIT);
+
+	public static MailDto mailDto = new MailDto("제목", new String[] {"naayen@naver.com"}, "body");
 }
